DPD test distro v04
CMake: disable not required features/modules by default
LMS API: add function for selecting chip
Do not enable TX AFE when streaming is started
FFTViewer uses LMS1 for TX (via DAC) and LMS2 for RX
Set sample rate only configures active chip
Add PA controls to board controls panel

+++ src/DPDTest/DPDTest.h
@@ -50,7 +50,7 @@ protected:
     
     void OnbtnCaptureClicked(wxCommandEvent& event);
     lms_device_t *lmsControl;
-    float mLmsNyquist_MHz;
+    double mLmsNyquist_MHz;
 
     void PlotFFT(OpenGLGraph* plot, const kiss_fft_cpx* fftOutput, const int samplesCount, const float nyquist_MHz);
     void PlotTimeDomain(OpenGLGraph* plot, const kiss_fft_cpx* samples, const int samplesCount);
